+++ cultivar-core/src/main/java/com/readytalk/cultivar/management/ShutdownManager.java
@@ -1,0 +1,11 @@
+package com.readytalk.cultivar.management;
+
+import com.google.common.annotations.Beta;
+
+/**
+ * Interface to abstract how to register a shutdown hook.
+ */
+@Beta
+public interface ShutdownManager {
+    void registerHook(final Runnable hook);
+}