+++ cultivar-discovery/src/main/java/com/readytalk/cultivar/discovery/RegistrationService.java
@@ -1,7 +1,11 @@
 package com.readytalk.cultivar.discovery;
 
+import javax.annotation.concurrent.ThreadSafe;
+
 /**
  * Registers a service with discovery. Typed to represent the different payloads.
  */
+@ThreadSafe
 public interface RegistrationService<T> extends DiscoveryService {
+
 }