+++ cultivar-discovery/src/test/java/com/readytalk/cultivar/discovery/DefaultRegistrationServiceTest.java
@@ -1,6 +1,8 @@
 package com.readytalk.cultivar.discovery;
 
+import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyZeroInteractions;
 
 import org.apache.curator.x.discovery.ServiceDiscovery;
 import org.apache.curator.x.discovery.ServiceInstance;
@@ -12,10 +14,10 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import com.google.common.base.VerifyException;
 import com.google.inject.util.Providers;
 
 @RunWith(MockitoJUnitRunner.class)
+@SuppressWarnings("unchecked")
 public class DefaultRegistrationServiceTest {
 
     @Rule
@@ -48,6 +50,17 @@ public class DefaultRegistrationServiceTest {
     }
 
     @Test
+    public void register_AlreadyRegistered_DoesNothing() throws Exception {
+        discoveryRegistration.register();
+
+        reset(discovery);
+
+        discoveryRegistration.register();
+
+        verifyZeroInteractions(discovery);
+    }
+
+    @Test
     public void shutDown_AfterStartUp_UnregistersService() throws Exception {
         discoveryRegistration.startUp();
 
@@ -57,9 +70,10 @@ public class DefaultRegistrationServiceTest {
     }
 
     @Test
-    public void shutDown_WithoutStartup_ThrowsVerifyException() throws Exception {
-        thrown.expect(VerifyException.class);
+    public void shutDown_WithoutStartup_DoesNothing() throws Exception {
 
         discoveryRegistration.shutDown();
+
+        verifyZeroInteractions(discovery);
     }
 }